Moved sensitive config info to the instance folder
Just need to remember to upload that folder to AWS.
