+++ .gitignore
@@ -1,8 +1,8 @@
-*.pyc
-
 # Elastic Beanstalk Files
 .elasticbeanstalk/*
 !.elasticbeanstalk/*.cfg.yml
 !.elasticbeanstalk/*.global.yml
 
 *.pyc
+
+instance/
