+++ hiuni_reports/data_helper.py
@@ -4,7 +4,7 @@
 
 
 
-from data_model import * # The data model & the database connection.
+from models import * # The data model & the database connection.
 import csv # For reading CSVs.
 from decorators import async
